+++ NEWS.md
@@ -1,6 +1,7 @@
 # spant 0.8.0
 * Updated plotting modes to be one of : "re", "im", "mod" or "arg".
 * Updated int_spec function to use "re", "im", or "mod".
+* Added a function to replicate data across a particular dimension.
 
 # spant 0.7.0
 * Added new function for frequency drift correction.